requirements.txt
astroid==2.2.5
cffi==1.17.1
click==7.1.2
cryptography==44.0.0
ecdsa==0.13.2
Flask==2.0.0
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.5.0
future==0.17.1
isort==4.3.18
itsdangerous==2.0.1
Jinja2==3.0.0
lazy-object-proxy==1.4.0
MarkupSafe==2.0.1
mccabe==0.6.1
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.3.1
pylint==2.3.1
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.12.0
SQLAlchemy==1.4.41
typed-ast==1.5.5
Werkzeug==2.0.0
wrapt==1.11.1
